BigQuery: move related code from AvroUtils -> BigQueryAvroUtils
